+++ assets/BoxPage.8047f68b.js
@@ -1,0 +1,20 @@
+var m=Object.defineProperty;var o=(e,a)=>m(e,"name",{value:a,configurable:!0});import{u as i,B as t}from"./index.694b1898.js";import{C as p}from"./ComponentInfo.cb73195c.js";import{a as n}from"./vendor.37d14b95.js";import"./Link.89a2d420.js";import"./copyText.f2e0ee2e.js";import"./PaperCode.464dde66.js";function r(){const e=i();return n(t,{get sx(){return{width:300,height:300,backgroundColor:e.palette.primary.dark,"&:hover":{backgroundColor:e.palette.primary.main,opacity:.9}}}})}o(r,"BoxSx");r.code=`import Box from "@suid/material/Box";
+import useTheme from "@suid/material/styles/useTheme";
+
+export default function BoxSx() {
+  const theme = useTheme();
+  return (
+    <Box
+      sx={{
+        width: 300,
+        height: 300,
+        backgroundColor: theme.palette.primary.dark,
+        "&:hover": {
+          backgroundColor: theme.palette.primary.main,
+          opacity: 0.9,
+        },
+      }}
+    />
+  );
+}
+`;function f(){return n(p,{get name(){return t.name},docsApiName:"components/box",moreExamples:!1,examples:[{component:r,title:!1}]})}o(f,"BoxPage");export{f as default};