+++ assets/LinksPage.a0b04023.js
@@ -1,0 +1,27 @@
+var a=Object.defineProperty;var r=(t,i)=>a(t,"name",{value:i,configurable:!0});import{L as e}from"./Link.89a2d420.js";import{C as l}from"./ComponentInfo.cb73195c.js";import{B as m}from"./index.694b1898.js";import{a as n}from"./vendor.37d14b95.js";import"./copyText.f2e0ee2e.js";import"./PaperCode.464dde66.js";function o(){return n(m,{sx:{typography:"body1","& > :not(style) + :not(style)":{ml:2}},onClick:t=>t.preventDefault(),get children(){return[n(e,{href:"#",target:"none",children:"Link"}),n(e,{href:"#",color:"inherit",target:"none",children:'color="inherit"'}),n(e,{href:"#",variant:"body2",target:"none",children:'variant="body2"'})]}})}r(o,"BasicLinks");o.code=`import Box from "@suid/material/Box";
+import Link from "@suid/material/Link";
+
+export default function BasicLinks() {
+  return (
+    <Box
+      sx={{
+        typography: "body1",
+        "& > :not(style) + :not(style)": {
+          ml: 2,
+        },
+      }}
+      onClick={(event) => event.preventDefault()}
+    >
+      <Link href="#" target="none">
+        Link
+      </Link>
+      <Link href="#" color="inherit" target="none">
+        {'color="inherit"'}
+      </Link>
+      <Link href="#" variant="body2" target="none">
+        {'variant="body2"'}
+      </Link>
+    </Box>
+  );
+}
+`;function L(){return n(l,{get name(){return e.name},docsName:"links",examples:[o]})}r(L,"LinksPage");export{L as default};