+++ assets/StackPage.1a83f334.js
@@ -1,0 +1,23 @@
+var c=Object.defineProperty;var n=(t,m)=>c(t,"name",{value:m,configurable:!0});import{S as a}from"./copyText.f2e0ee2e.js";import{C as i}from"./ComponentInfo.cb73195c.js";import{r as p,a as e,t as s}from"./vendor.37d14b95.js";import{s as d,P as l}from"./index.694b1898.js";import"./Link.89a2d420.js";import"./PaperCode.464dde66.js";const u=s("<div></div>"),r=d(l)(({theme:t})=>({...t.typography.body2,padding:t.spacing(1),textAlign:"center",color:t.palette.text.secondary}));function o(){return(()=>{const t=u.cloneNode(!0);return p(t,e(a,{spacing:2,get children(){return[e(r,{children:"Item 1"}),e(r,{children:"Item 2"}),e(r,{children:"Item 3"})]}})),t})()}n(o,"BasicStack");o.code=`import Paper from "@suid/material/Paper";
+import Stack from "@suid/material/Stack";
+import styled from "@suid/material/styles/styled";
+
+const Item = styled(Paper)(({ theme }) => ({
+  ...theme.typography.body2,
+  padding: theme.spacing(1),
+  textAlign: "center",
+  color: theme.palette.text.secondary,
+}));
+
+export default function BasicStack() {
+  return (
+    <div>
+      <Stack spacing={2}>
+        <Item>Item 1</Item>
+        <Item>Item 2</Item>
+        <Item>Item 3</Item>
+      </Stack>
+    </div>
+  );
+}
+`;function x(){return e(i,{get name(){return a.name},examples:[{bgcolor:"contrasted",component:o}]})}n(x,"StackPage");export{x as default};