+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # @suid/site
 
+## 0.0.22
+
+### Patch Changes
+
+- [`ef21396`](https://github.com/swordev/suid/commit/ef21396ead087c64b44a1744834f09a68a1f45a2) Thanks [@juanrgm](https://github.com/juanrgm)! - Add package description
+
+- Updated dependencies [[`ef21396`](https://github.com/swordev/suid/commit/ef21396ead087c64b44a1744834f09a68a1f45a2)]:
+  - @suid/base@0.0.15
+  - @suid/icons-material@0.0.16
+  - @suid/material@0.0.16
+  - @suid/system@0.0.14
+  - @suid/types@0.0.6
+  - @suid/utils@0.0.7
+
 ## 0.0.21
 
 ### Patch Changes